Haunted Memories v.7.0.3
Another Genius moment of me.
When trying to use an object from Inventory it would work when you are near it. only if it "overlays" you and you are forcefully touching it.
But as the genius as I am it is fixed again.

NEXT: F3 Finish Rooms and Objects (+ riddles maybe)